+++ server/routes/upload-photo.route.js
@@ -9,6 +9,7 @@ const config = require('../utils/config')
 const RedisService = require('../services/redis.service')
 const { Paths, RedisKeys, Views } = require('../utils/constants')
 const { buildErrorSummary } = require('../utils/validation')
+const { checkForDuplicates, checkForFileSizeError } = require('../utils/upload')
 
 const MAX_PHOTOS = 6
 const MAX_FILES_IN_REQUEST_PAYLOAD = 1
@@ -29,7 +30,10 @@ const handlers = {
       return h.redirect(Paths.YOUR_PHOTOS)
     }
 
-    const errors = await _checkForFileSizeError(request)
+    const errors = await checkForFileSizeError(
+      request,
+      RedisKeys.UPLOAD_PHOTO_ERROR
+    )
 
     return h.view(Views.UPLOAD_PHOTO, {
       ...(await _getContext(request)),
@@ -172,46 +176,12 @@ const _validateForm = (payload, uploadData) => {
       name: 'files',
       text: 'The file must be a JPG or PNG'
     })
-  } else if (_checkForDuplicates(payload, uploadData)) {
+  } else if (checkForDuplicates(payload, uploadData)) {
     errors.push({
       name: 'files',
       text: "You've already uploaded that image. Choose a different one"
     })
   }
-
-  return errors
-}
-
-const _checkForDuplicates = (payload, uploadData) => {
-  let duplicateFound
-
-  if (uploadData.files && uploadData.fileSizes) {
-    for (let i = 0; i < uploadData.files.length; i++) {
-      if (
-        uploadData.files[i] === payload.files.filename &&
-        uploadData.fileSizes[i] === payload.files.bytes
-      ) {
-        duplicateFound = true
-        break
-      }
-    }
-  }
-
-  return duplicateFound
-}
-
-const _checkForFileSizeError = async request => {
-  const errors = []
-  if (
-    (await RedisService.get(request, RedisKeys.UPLOAD_PHOTO_ERROR)) === 'true'
-  ) {
-    errors.push({
-      name: 'files',
-      text: `The file must be smaller than ${config.maximumFileSize}mb`
-    })
-  }
-
-  await RedisService.set(request, RedisKeys.UPLOAD_PHOTO_ERROR, false)
 
   return errors
 }
